@@ -10,32 +10,35 @@
     "start": "nuxt start",
     "generate": "nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "lintfix": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
+    "lint:fix": "eslint --fix --ext .js,.vue --ignore-path .gitignore .",
     "precommit": "npm run lint"
   },
   "dependencies": {
-    "@nuxtjs/axios": "^5.3.6",
+    "@fortawesome/fontawesome-svg-core": "^1.2.17",
+    "@fortawesome/free-brands-svg-icons": "^5.8.1",
+    "@fortawesome/free-solid-svg-icons": "^5.8.1",
+    "@fortawesome/vue-fontawesome": "^0.1.6",
+    "@nuxtjs/axios": "^5.4.1",
     "@nuxtjs/dotenv": "^1.3.0",
-    "@nuxtjs/google-analytics": "^2.0.2",
+    "@nuxtjs/google-analytics": "^2.2.0",
     "@nuxtjs/markdownit": "^1.2.3",
     "@nuxtjs/pwa": "^2.6.0",
-    "@nuxtjs/sentry": "^2.2.1",
-    "@nuxtjs/sitemap": "^0.2.0",
+    "@nuxtjs/sentry": "^2.3.1",
+    "@nuxtjs/sitemap": "^0.2.2",
     "cross-env": "^5.2.0",
     "markdown-it-highlightjs": "^3.0.0",
-    "nuxt": "^2.4.3",
-    "nuxt-buefy": "^0.3.3",
-    "nuxt-i18n": "^5.8.0",
-    "nuxt-matomo": "^1.2.1",
-    "nuxt-robots-module": "^1.4.0",
+    "nuxt": "^2.5.1",
+    "nuxt-buefy": "^0.3.4",
+    "nuxt-i18n": "^5.8.5",
+    "nuxt-matomo": "^1.2.2",
+    "nuxt-robots-module": "^1.5.1",
     "nuxt-webfontloader": "^1.0.0",
     "vue-cookie-accept-decline": "^5.1.0",
     "vue-recaptcha": "^1.1.1"
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.7.1",
     "babel-eslint": "^8.2.1",
-    "eslint": "^5.13.0",
+    "eslint": "^5.15.3",
     "eslint-config-prettier": "^3.6.0",
     "eslint-loader": "^2.1.2",
     "eslint-plugin-prettier": "2.6.2",
@@ -43,7 +46,7 @@
     "glob": "^7.1.3",
     "husky": "^1.3.1",
     "node-sass": "^4.11.0",
-    "nodemon": "^1.18.9",
+    "nodemon": "^1.18.10",
     "prettier": "1.14.3",
     "sass-loader": "^7.1.0",
     "shrink-ray-current": "^3.0.2"
