@@ -8,17 +8,17 @@ module.exports = {
     parser: 'babel-eslint'
   },
   extends: [
+    // '@nuxtjs',
     'plugin:vue/recommended',
     'plugin:prettier/recommended'
   ],
-  // required to lint *.vue files
   plugins: [
     'vue',
     'prettier'
   ],
   // add your custom rules here
   rules: {
-    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
-    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
+    'no-console': process.env.NODE_ENV === 'production' ? 2 : 0,
+    'no-debugger': process.env.NODE_ENV === 'production' ? 2 : 0
   }
 }
